requirements.txt
pybullet>=3.2.5
gymnasium>=0.28.1
stable-baselines3>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
torch>=2.0.0
trimesh>=3.20.0
PyYAML>=6.0
